Refactor event handling into changelist
Basically I realized I need to properly handle only sending in changes
to the events we want to register and receive, making sure to delete
events for files we detect no longer exist (and for fd's we've
closed).

Added in some file ignore patterns, too, to prevent the issue with
emacs temp files (`.#*`) making things puke since technically I
believe they are symlinks elsewhere and we're not following symlinks
with `os.fwalk()` (yet).